Added a base class to hande Input & Output for list-item, list-item-swipe & list-item-menu, as mosst of them are the same and maintaining the interface between them are important as most of the data from list-item is simply passed through to swipe & menu.
Moved styling.
Added types file to share BoundaryClass and device types
