+++ libs/designsystem/list/src/list-item/list-item-menu/list-item-menu.component.ts
@@ -1,13 +1,8 @@
 import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'kirby-list-item-menu',
-  standalone: true,
-  imports: [CommonModule],
   templateUrl: './list-item-menu.component.html',
-  styleUrls: ['./list-item-menu.component.css']
+  styleUrls: ['./list-item-menu.component.css'],
 })
-export class ListItemMenuComponent {
-
-}
+export class ListItemMenuComponent {}
