+++ libs/designsystem/list/src/list-item/list-item-swipe/list-item-swipe.component.ts
@@ -1,13 +1,24 @@
-import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { AfterViewInit, Component, OnInit } from '@angular/core';
+import { PlatformService } from '@kirbydesign/designsystem/helpers';
+import ListItemBaseComponent from '../list-item.base.component';
 
 @Component({
   selector: 'kirby-list-item-swipe',
-  standalone: true,
-  imports: [CommonModule],
   templateUrl: './list-item-swipe.component.html',
-  styleUrls: ['./list-item-swipe.component.css']
+  styleUrls: ['../../list.component.scss', './list-item-swipe.component.scss'],
 })
-export class ListItemSwipeComponent {
+export class ListItemSwipeComponent extends ListItemBaseComponent implements OnInit, AfterViewInit {
+  constructor(_platform: PlatformService) {
+    super(_platform);
+  }
 
+  ngOnInit() {
+    this.initializeSwipeActions();
+  }
+
+  ngAfterViewInit(): void {
+    if (!this.itemTemplate) {
+      console.warn('No item template was provided.');
+    }
+  }
 }
