--- libs/designsystem/list/src/list-item/list-item.component.scss
@@ -1,13 +1,0 @@
-@use '@kirbydesign/core/src/scss/utils';
-
-$divider-color: utils.get-color('background-color');
-
-:host {
-  overflow: hidden;
-}
-
-:host-context(.has-divider) {
-  ion-item-sliding:not(.last) {
-    border-bottom: 1px solid $divider-color;
-  }
-}
