+++ libs/designsystem/list/src/list-item/list-item.component.ts
@@ -1,39 +1,18 @@
-import {
-  AfterViewInit,
-  Component,
-  EventEmitter,
-  Input,
-  OnInit,
-  Output,
-  TemplateRef,
-  ViewChild,
-} from '@angular/core';
-import { IonItemSliding } from '@ionic/angular';
-
-import { ThemeColor } from '@kirbydesign/core';
+import { Component, Input } from '@angular/core';
 import { PlatformService } from '@kirbydesign/designsystem/helpers';
-
-import { ListSwipeAction, ListSwipeDirection, ListSwipeEnd } from '../list-swipe-action.type';
 import ListItemBaseComponent from './list-item.base.component';
-
-export type BoundaryClass = 'first' | 'last';
+import { BoundaryClass, Device } from './list-item.types';
 
 @Component({
   selector: 'kirby-list-item',
   templateUrl: './list-item.component.html',
-  styleUrls: ['../list.component.scss', './list-item.component.scss'],
+  styleUrls: ['./list-item.component.scss'],
 })
-export class ListItemComponent extends ListItemBaseComponent implements OnInit, AfterViewInit {
+export class ListItemComponent extends ListItemBaseComponent {
+  @Input() device: Device;
+  @Input() boundaryClass: BoundaryClass;
+
   constructor(private _platform: PlatformService) {
     super(_platform);
-  }
-  ngOnInit() {
-    this.initializeSwipeActions();
-  }
-
-  ngAfterViewInit(): void {
-    if (!this.itemTemplate) {
-      console.warn('No item template was provided.');
-    }
   }
 }
