+++ libs/designsystem/list/src/list-item/list-item.types.ts
@@ -1,0 +1,8 @@
+/**
+ * Detmines if the item is the first or last in the list.
+ */
+export type BoundaryClass = 'first' | 'last';
+/**
+ * Determines if the list is rendered on a touch device or a desktop.
+ */
+export type Device = 'touch' | 'desktop';
