+++ libs/designsystem/list/src/list.component.scss
@@ -169,6 +169,7 @@ ion-item-options[side='end'] {
 
   ion-item.first,
   ion-item-sliding.first,
+  kirby-item.first,
   ion-list-header {
     border-top-left-radius: utils.$border-radius;
     border-top-right-radius: utils.$border-radius;
@@ -187,6 +188,7 @@ ion-item-options[side='end'] {
 
   ion-item.last,
   ion-item-sliding.last,
+  kirby-item.last,
   .footer {
     border-bottom-left-radius: utils.$border-radius;
     border-bottom-right-radius: utils.$border-radius;
